add request for add\del data with database for methods form

+++ utils_db/services.py
@@ -1,4 +1,4 @@
-from pony.orm import db_session, TransactionError, select
+from pony.orm import db_session, TransactionError
 from abc import ABC, abstractmethod
 
 from utils_db.Db_models import ConnDb
@@ -18,7 +18,7 @@ class AbstractService(ABC):
 
 class get_instance(AbstractService):
 
-    def __init__(self, name_db, incoming_data, value_for_update=None):
+    def __init__(self, name_db: str, incoming_data=None, value_for_update=None):
         self.name_db = name_db
         self.db = ConnDb()
         self.incoming_data = incoming_data
@@ -43,10 +43,37 @@ class get_instance(AbstractService):
         with db_session:
             return self.initial.get(**self.value_for_update['search_criteria'])
 
+    def _check_uniqueness(self):
+        if self.value_for_update:
+            self.get_record_by_criteria()
+
+            with db_session:
+
+                if self.incoming_data:
+                    if self.initial.get(Tcn=self.incoming_data.get('Fuse').get('Tcn')) is not None:
+                        return f"Запись  {self.incoming_data['Fuse']['Tcn']} уже существует."
+                    if self.initial.get(Name=self.incoming_data.get('Name')) is not None:
+                        return f"Запись  {self.incoming_data['Name']} уже существует."
+                if self.value_for_update:
+                    self._delete_entity()
+
+        return None
+
     def _create_entity(self):
+        uniqueness_error = self._check_uniqueness()
+        if uniqueness_error:
+            return uniqueness_error
+
         try:
             with db_session:
-                return self.initial(**self.incoming_data)
+                if self.incoming_data.get('Fuse') is not None:
+
+                    self.name_db = 'FuseTarget'
+                    fuse_target = self.initial_entity()
+                    self.incoming_data['FuseTarget']['Fuse'] = self.initial(**self.incoming_data['Fuse'])
+                    fuse_target(**self.incoming_data['FuseTarget'])
+                else:
+                    self.initial(**self.incoming_data)
         except AttributeError as e:
             print("Ошибка доступа к атрибуту:", e)
         except TypeError as e:
@@ -75,7 +102,9 @@ class get_instance(AbstractService):
     def _delete_entity(self):
         try:
             with db_session:
-                return self.get_record_by_criteria().delete()
+                if self.initial.get(Name=self.value_for_update.get('search_criteria').get('Name')) is not None:
+                    return self.get_record_by_criteria().delete()
+
         except AttributeError as e:
             print("Ошибка доступа к атрибуту:", e)
         except TypeError as e:
@@ -83,6 +112,5 @@ class get_instance(AbstractService):
         except Exception as e:
             print("Произошла ошибка:", e)
         return False
-
